Move signup and login frontend to shared folder
